[Glitch] Rewrite AnimatedNumber component with React hooks
Port ab740f464a8e5aa6b5f78c0ddab3c8e18698d810 to glitch-soc
